+++ src/writer/mod.rs
@@ -1,6 +1,7 @@
 //! Writers for spreadsheet file formats.
 
 mod chart;
+mod comment;
 pub mod csv;
 pub mod html;
 mod image;